rdr: add custom ip and device modifier
This is enhancement for the rdr command to allow to specify
the target host ip and network device.

The "ip" modifier sets the target ip replacing the "to any" to "to <IP>".
The "dev" modifier replaces the default device to the specified one.

Both are stored to the persistent rdr.conf.
The format of old rdr.conf entries does not change.

All of these rule combinations are now supported (any has a valid usecase):
tcp 22 22
ip 192.168.1.1 tcp 22 22
dev wg0 tcp 22 22
ip 192.168.1.1 dev wg0 tcp 22 22
ip 192.168.1.1 dev wg0 tcp 23 22 log