add support for static mac address for jails
This commit will generate a static MAC address for each jail, based on the name of the jail. It will use the first half (xx:xx:xx) of the host MAC to avoid network clashes, and generate a random HEX string from the hashed name of the jail. It will then add that random 5 character HEX string in MAC format, and add an "a" and "b" for the host and jail respectively. 

This way a jail can retain it's MAC ID even if it is deleted and reinstalled, as long as the same name is retained.