bastille: Support user based config
This PR will allow bastille to be configured based on the user you run as.

To load the config for a specifig user, you must be logged in as the specified user, and there must be a file called "bastille_user.conf" inside /usr/local/etc/bastille or else bastille will just load the default config file.