Implement parallel mode
This PR allows to run commands on multiple jails in parallel. Simply use '-p 3' to set a process limit.

'bastille -p 3 start all' will start all jails but run 3 processes at a time.

This is a very rough implementation, and could be more streamlined, but here it is to test.