list: properly list jail type
@michael-o 

This is actually a better solution. It didn't cross my mind that we are actually setting IS_FREEBSD_JAIL and IS_LINUX_JAIL in the list command.

For now, if both are detected, it will set JAIL_TYPE to "hybrid" but if that is not agreeable, we can simply leave it at "thick".